replaced QtCore.QCoreApplication.processEvents()
