+++ tests/_test_msui/test_linearview.py
@@ -31,7 +31,7 @@ from PyQt5 import QtTest, QtCore
 from mslib.msui import flighttrack as ft
 import mslib.msui.linearview as tv
 from mslib.msui.msui import MSUIMainWindow
-from mslib.msui.mpl_qtwidget import _DEFAULT_SETTINGS_LINEARVIEW
+from mslib.msui.viewplotter import _DEFAULT_SETTINGS_LINEARVIEW
 
 
 class Test_MSS_LV_Options_Dialog:
